CLI arguments introduced.
Well ... ok build via CLI was a pain. But should still be a viable option. CLI arguments seem to be way simpler to use. Ml-agents gives via --env-args arguments to the unity build. Obvious problems when running in the editor. More comments needed. More logic needs to be added to the python script.